ugly but working

+++ server/routes/api.js
@@ -1,6 +1,7 @@
 const router = require('express').Router()
 
 const User = require('./../models/user')
+const Poll = require('./../models/poll')
 // const CONFIG = require('./../config')
 const my = require('./../helper')
 const CustomError = my.CustomError
@@ -90,29 +91,50 @@ router.delete('/deleteUser/:id', my.verifyToken, my.AdminGuard, (req, res, next)
   //   })
 })
 
-const Poll = require('./../models/poll')
+router.get('/get-all-polls/:id?', (req, res, next) => {
+  const query = {}
+  if (req.params.id) {
+    query.owner = req.params.id
+  }
+  Poll.find(query, '_id title', (err, polls) => {
+    if (err) {
+      return next(err)
+    }
+    res.json(polls)
+  })
+})
 
 router.get('/get-poll/:id', (req, res, next) => {
   Poll.findById(req.params.id).exec()
     .then(poll => {
-      res.json(poll)
+      if (!poll) {
+        throw new CustomError('No Poll Found', 404)
+      }
+      req.poll = poll
+      return User.findById(poll.owner).exec()
+      // res.json(poll)
+    })
+    .then(user => {
+      req.poll.owner = user.username || '*Unknown*'
+      res.json(req.poll)
     })
     .catch(err => {
       return next(err)
     })
 })
 
-router.post('/make-poll', my.verifyToken, pollCheck, (req, res, next) => {
+router.post('/make-poll', my.verifyToken, my.UserGuard, pollCheck, (req, res, next) => {
   const newPoll = new Poll()
   newPoll.title = req.body.title
-  newPoll.owner = req.body.owner
-  newPoll.options = req.body.options
-  newPoll.votes = req.body.votes
+  newPoll.owner = req.user._id
+  newPoll.labels = req.body.options
+  newPoll.data = new Array(req.body.options.length).fill(0)
   newPoll.save(err => {
     if (err) {
       return next(err)
     }
-    res.send(newPoll._id)
+    console.log(newPoll._id)
+    res.json({ id: newPoll._id })
   })
 })
 
@@ -120,26 +142,58 @@ function pollCheck (req, res, next) {
   if (!req.body.title) {
     return next(new CustomError('Missing Title', 400))
   }
-  if (!req.body.owner) {
-    return next(new CustomError('Missing Owner', 400))
-  }
+  // if (!req.body.owner) {
+  //   return next(new CustomError('Missing Owner', 400))
+  // }
   if (!req.body.options || !req.body.options.length) {
     return next(new CustomError('Missing Options', 400))
   }
-  if (!req.body.votes || !req.body.options.votes) {
-    return next(new CustomError('Missing Votes', 400))
-  }
-  next()
+  // if (!req.body.votes || !req.body.options.votes) {
+  //   return next(new CustomError('Missing Votes', 400))
+  // }
+  Poll.findOne({ title: req.body.title }).exec()
+    .then(poll => {
+      if (poll) {
+        return next(new CustomError('Poll With Same Title Already Exists', 409))
+      }
+      return next()
+    })
+    .catch(err => {
+      next(err)
+    })
 }
 
-router.post('/poll-vote/:id', my.verifyToken, (req, res, next) => {
+router.put('/poll-vote/:id', (req, res, next) => {
   Poll.findById(req.params.id).exec()
     .then(poll => {
-      const i = poll.options.indexOf(req.body.choice)
-      if (i < 0) {
-        throw new CustomError('Invalid Choice', 403)
+      const i = req.body.index
+      if (i < 0 || i >= poll.labels.length) {
+        throw new CustomError('Invalid Choice', 401)
       }
-      poll.votes[i]++
+      // poll.data[i]++
+      const update = {}
+      const entry = {}
+      entry['data.' + i] = 1
+      update.$inc = entry
+      return Poll.findByIdAndUpdate(poll._id, update).exec()
+      // return poll.save()
+    })
+    .then(poll => {
+      res.send('Poll Updated')
+    })
+    .catch(err => {
+      return next(err)
+    })
+})
+
+router.put('/poll-add-option/:id', my.verifyToken, (req, res, next) => {
+  Poll.findById(req.params.id).exec()
+    .then(poll => {
+      if (poll.labels.length > 9) {
+        throw new CustomError('Maximum of 10 options', 401)
+      }
+      poll.labels.push(req.body.newOption)
+      poll.data.push(0)
       return poll.save()
     })
     .then(poll => {
@@ -150,27 +204,15 @@ router.post('/poll-vote/:id', my.verifyToken, (req, res, next) => {
     })
 })
 
-router.post('/poll-add-option/:id', my.verifyToken, (req, res, next) => {
+router.delete('/delete-poll/:id', my.verifyToken, my.UserGuard, (req, res, next) => {
   Poll.findById(req.params.id).exec()
     .then(poll => {
-      if (poll.options.length > 9) {
-        throw new CustomError('Maximum of 10 options', 403)
+      if (poll.owner !== req.user._id.toString() && req.user.role !== 'admin') {
+        throw new CustomError(`Only the owner or an Admin can delete Poll ID: ${req.params.id}`, 401)
       }
-      poll.options.push(req.body.newOption)
-      poll.votes.push(0)
-      return poll.save()
+      return Poll.findByIdAndRemove(req.params.id).exec()
     })
-    .then(poll => {
-      res.send('Poll Updated')
-    })
-    .catch(err => {
-      return next(err)
-    })
-})
-
-router.delete('/delete-poll/:id', my.verifyToken, (req, res, next) => {
-  Poll.findByIdAndRemove(req.params.id).exec()
-    .then(poll => {
+    .then(result => {
       res.send(`Poll ID: ${req.params.id} Deleted`)
     })
     .catch(err => {
